Update unloadability snippets
Reduces the number of snippet files needed using the snippet tags
instead of separate file for each snippet.
Moves the files to snippets/standard/assembly/unloading.
