--- snippets/core/tutorials/unloading/create_test_assemblyloadcontext.cs
@@ -1,2 +1,0 @@
-var alc = new TestAssemblyLoadContext();
-Assembly a = alc.LoadFromAssemblyPath("absolute/path/to/your/assembly");
