--- snippets/core/tutorials/unloading/execute_method_in_assemblyloadcontext.cs
@@ -1,2 +1,0 @@
-var args = new object[1] {new string[] {"Hello"}};
-int result = (int)a.EntryPoint.Invoke(null, args);
