--- snippets/core/tutorials/unloading/simple_assemblyloadcontext.cs
@@ -1,11 +1,0 @@
-class TestAssemblyLoadContext : AssemblyLoadContext
-{
-    public TestAssemblyLoadContext() : base(isCollectible: true)
-    {
-    }
-
-    protected override Assembly Load(AssemblyName name)
-    {
-        return null;
-    }
-}
