--- snippets/core/tutorials/unloading/unload_assemblyloadcontext.cs
@@ -1,1 +1,0 @@
-alc.Unload();
