--- snippets/core/tutorials/unloading/wait_for_unload_completion.cs
@@ -1,5 +1,0 @@
-for (int i = 0; testAlcWeakRef.IsAlive && (i < 10); i++)
-{
-    GC.Collect();
-    GC.WaitForPendingFinalizers();
-}
